Projects/util.py
import boto3
import catboost
import causalml
import diagrams
import dotenv
import duckdb
import emoji
import fuzzywuzzy
import google
import google.api_core
import google.cloud
import google.cloud.discoveryengine
import google.cloud.storage
import gower
import hdbscan
import jinja2
import jinjasql
import jupyter
import keyring
import matplotlib
import msal
import nltk
import numpy
import openai
import openpyxl
import optuna
import pandas
import pandasql
import pillow
import plotly
import pyjwt
import pyperclip
import python_dateutil
import dotenv as python_dotenv
import pptx as python_pptx
import rapidfuzz
import ratelimit
import rpy2
import sklearn
import skopt
import scipy
import seaborn
import selenium
import snowflake.connector
import spacy
import sqlalchemy
import statsmodels
import tiktoken
import tomli
import tqdm
import umap
import webdriver_manager
import wordcloud
import xgboost
import xlsxwriter
import xlwings
import yake

# Utility function to demonstrate usage of all imported packages
def demonstrate_packages():
    print("Demonstrating usage of imported packages:")

    # Example usage of numpy
    arr = numpy.array([1, 2, 3])
    print("Numpy array:", arr)

    # Example usage of pandas
    df = pandas.DataFrame({'A': [1, 2], 'B': [3, 4]})
    print("Pandas DataFrame:", df)

    # Example usage of matplotlib
    matplotlib.pyplot.plot([1, 2, 3], [4, 5, 6])
    matplotlib.pyplot.title("Example Plot")
    matplotlib.pyplot.show()

    # Example usage of requests (if applicable)
    # response = requests.get("https://api.example.com")
    # print("Requests response:", response.status_code)

    # Add more examples as needed for other packages

if __name__ == "__main__":
    demonstrate_packages()