pyproject.toml
[tool.poetry]

name = "RNA-Project"

version = "0.1.0"

description = ""

authors = [

    "Colin Farmer <colinjfarmer@gmail.com>",

    "Ethan Bouvet <ebouve@uw.edu>",

]

readme = "README.md"


 [tool.poetry.dependencies]

python = "^3.9,<3.11"

boto3 = "^1.28.41"

catboost = "^1.2.7"

causalml = "^0.15.3"

diagrams = "^0.23.3"

dotenv = "^0.9.9"

duckdb = "0.9.2"

emoji = "^2.8.0"

fuzzywuzzy = "^0.18.0"

google = "^3.0.0"

google-api-core = "^2.16.1"

google-cloud = "^0.34.0"

google-cloud-discoveryengine = "^0.11.6"

google-cloud-storage = "^2.14.0"

gower = "^0.1.2"

hdbscan = "^0.8.33"

jinja2 = "^3.1.2"

jinjasql = "^0.1.8"

jupyter = "^1.0.0"

keyring = "^24.2.0"

matplotlib = "^3.7.2"

msal = "^1.24.0"

nltk = "^3.8.1"

numpy = "^1.25.2"

openai = "^0.28.0"

openpyxl = "^3.1.2"

optuna = "^4.2.1"

pandas = "^2.0.3"

pandasql = "^0.7.3"

pillow = "^9.5"

plotly = "^5.16.1"

pyjwt = "^2.8.0"

pyperclip = "^1.9.0"

python-dateutil = "^2.8.2"

python-dotenv = "^1.0.1"

python-pptx = "^0.6.22"

rapidfuzz = "^3.11.0"

ratelimit = "^2.2.1"

rpy2 = "^3.5.17"

scikit-learn = "^1.3.0"

scikit-optimize = "^0.10.2"

scipy = "^1.11.2"

seaborn = "0.13.1"

selenium = "^4.30.0"

snowflake-connector-python = {extras = ["pandas", "secure-local-storage"], version = "^3.2.0"}

spacy = "3.8.3"

sqlalchemy = "^2.0.20"

statsmodels = "^0.14.0"

tiktoken = "^0.7.0"

tomli = "^2.0.1"

tqdm = "^4.66.1"

umap-learn = "^0.5.4"

webdriver-manager = "^4.0.2"

wordcloud = "^1.9.2"

xgboost = "^2.0.0"

xlsxwriter = "^3.1.2"

xlwings = "^0.30.11"

yake = "^0.4.8"

 

[tool.poetry.group.dev.dependencies]

pyclean = "^2.7.4"

ruff = "^0.1.8"

sqlfluff = "^2.3.2"

 

 

[tool.poetry.group.umap-learn.dependencies]

numba = "^0.58.0"

 

[build-system]

requires = ["poetry-core"]

build-backend = "poetry.core.masonry.api"

 

[tool.ruff]

extend-include = ["*.ipynb"]  # jupyter

extend-select = ["I"]  # isort

target-version = "py39"

select = ["D"]  # docs

#ignore = ["D417"]

 

[tool.ruff.pydocstyle]

convention = "google"

 

[tool.ruff.mccabe]

max-complexity = 10